+++ grupo4/src/main/java/com/trabalhofinal/grupo4/services/ClienteServices.java
@@ -3,7 +3,6 @@ package com.trabalhofinal.grupo4.services;
 
 import java.util.List;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -36,8 +35,23 @@ public class ClienteServices {
 		return clienteRepo.save(cliente);
 	}
 	
-	public void deletarCliente(Cliente cliente) {
-		clienteRepo.delete(cliente);
-		
-	}
+	public Boolean deletarCliente(Cliente cliente) {
+        if(cliente == null)
+            return false;
+
+        Cliente clienteExistente = buscarClientePorId(cliente.getIdCliente());
+
+        if(clienteExistente == null)
+            return false;
+
+        clienteRepo.delete(cliente);
+
+        Cliente clienteContinuaExistindo = buscarClientePorId(cliente.getIdCliente());
+
+        if(clienteContinuaExistindo == null)
+            return true;
+
+        return false;
+
+    }
 }
